update initial setup to handle telemetry consent
